+++ include/reverse/boolean_engine/logical_query.hpp
@@ -83,7 +83,7 @@ template <typename Schema>
 void LogicalQuery<Schema>::and_or_add_subnode(
         const ExecutorNode<Schema>& node,
         OperatorBooleanExecutor<Schema>* result) {
-    for (int i = 0; i < node._sub_nodes.size(); ++i) {
+    for (size_t i = 0; i < node._sub_nodes.size(); ++i) {
         const ExecutorNode<Schema>* sub_node = node._sub_nodes[i];
         BooleanExecutor<Schema> *tmp= parse_executor_node(*sub_node);
         if (tmp) {
@@ -104,7 +104,7 @@ void LogicalQuery<Schema>::weight_add_subnode(
     if (tmp) {
         weight_result->add_must(tmp);
     }
-    for (int i = 1; i < node._sub_nodes.size(); ++i) {
+    for (size_t i = 1; i < node._sub_nodes.size(); ++i) {
         const ExecutorNode<Schema>* sub_node = node._sub_nodes[i];
         BooleanExecutor<Schema> *tmp= parse_executor_node(*sub_node);
         if (tmp) {
